CentOS 7 gdb-rhbz795424-bitpos-23of25.patch: [PATCH 4/4] Bitpos expansion - tdep changes
http://sourceware.org/ml/gdb-patches/2012-09/msg00630.html
Subject: [PATCH 4/4] Bitpos expansion - tdep changes

--MP_/X_WjDOvz/B_fvlsrmCwRdxe
Content-Type: text/plain; charset=US-ASCII
Content-Transfer-Encoding: 7bit
Content-Disposition: inline

Hi,

This is the last part of the bitpos change.  This patch does the same
thing that patch 1/4 does, except that it does so in the *-tdep files
wherever necessary.  I separated these changes mainly because they are
related and can be reviewed independently.  Tested on Fedora 16 x86_64.

Regards,
Siddhesh

--MP_/X_WjDOvz/B_fvlsrmCwRdxe
Content-Type: text/plain
Content-Transfer-Encoding: quoted-printable
Content-Disposition: attachment; filename=ChangeLog-tdep

gdb/ChangeLog

	* alpha-tdep.c (alpha_push_dummy_call): Expand ACCUMULATE_SIZE,
	REQUIRED_ARG_REGS, OFFSET, LEN, TLEN to ssize_t.
	(struct alpha_arg): Expand members LEN, OFFSET to ssize_t.
	* amd64-tdep.c (amd64_push_arguments): Expand NUM_ELEMENTS,
	ELEMENT and LEN to LONGEST.
	(amd64_return_value): Expand LEN to LONGEST.
	* amd64-windows-tdep.c (amd64_windows_return_value): Expand LEN
	to LONGEST.
	* arm-tdep.c (arm_vfp_cprc_sub_candidate): Return LONGEST.
	Expand COUNT, SUB_COUNT to LONGEST.
	(arm_vfp_call_candidate): Expand C to LONGEST.
	(arm_push_dummy_call): Expand LEN to LONGEST.
	* avr-tdep.c (struct stack_item): Expand member LEN to ssize_t.
	(push_stack_item): Expand parameter LEN to ssize_t.
	(avr_push_dummy_call): Expand LAST_REGNUM, J, LEN to ssize_t.
	* bfin-tdep.c (bfin_push_dummy_call): Expand TOTAL_LEN,
	CONTAINER_LEN to ssize_t.
	* cris-tdep.c (struct stack_item): Expand member LEN to ssize_t.
	(push_stack_item): Expand parameter LEN to ssize_t.
	(cris_push_dummy_call): Expand LEN, REG_DEMAND, I to ssize_t.
	* h8300-tdep.c (h8300_push_dummy_call): Expand STACK_ALLOC,
	STACK_OFFSET to LONGEST.  Expand LEN, PADDED_LEN, OFFSET to
	ssize_t.
	* hppa-tdep.c (hppa64_push_dummy_call): Expand LEN to LONGEST.
	(hppa64_return_value): Likewise.
	* i386-darwin-tdep.c (i386_darwin_push_dummy_call): Expand
	ARGS_SPACE to LONGEST.
	* i386-tdep.c (i386_push_dummy_call): Expand ARG_SPACE,
	ARG_SPACE_USED, LEN to LONGEST.
	(i386_reg_struct_return_p): Expand LEN to LONGEST.
	(i386_convert_register_p): Likewise.
	(i386_register_to_value): Likewise.
	(i386_value_to_register): Likewise.
	* ia64-tdep.c (ia64_push_dummy_call): Expand ARGOFFSET, LEN,
	NSLOTS, MEMSLOTS to LONGEST.
	* iq2000-tdep.c (iq2000_push_dummy_call): Expand TYPELEN,
	STACKSPACE to LONGEST.
	* m32r-tdep.c (m32r_push_dummy_call): Expand LEN to LONGEST.
	* m68k-tdep.c (m68k_reg_struct_return_p): Expand LEN to LONGEST.
	(m68k_push_dummy_call): Expand LEN, CONTAINER_LEN, OFFSET to
	LONGEST.
	* m88k-tdep.c (m88k_store_arguments): Expand NUM_STACK_WORDS,
	LEN, STACK_WORD to LONGEST.
	* mep-tdep.c (push_large_arguments): Expand ARG_LEN to ULONGEST.
	* microblaze-tdep.c (microblaze_store_return_value): Expand LEN to
	LONGEST.
	* mips-tdep.c (mips_xfer_register): Expand parameter BUF_OFFSET to
	LONGEST.  Use plongest to format print BUF_OFFSET.
	(mips_eabi_push_dummy_call): Expand LEN to LONGEST.  Use plongest
	to format print LEN.
	(mips_n32n64_fp_arg_chunk_p): Expand parameter OFFSET to LONGEST.
	Expand POS to LONGEST.
	(mips_n32n64_push_dummy_call): Expand LEN to LONGEST.
	(mips_n32n64_return_value): Expand OFFSET to LONGEST.  Use
	plongest to format print OFFSET.
	(mips_o32_push_dummy_call): Expand LEN to LONGEST.  Use plongest
	to format print LEN.
	(mips_o64_push_dummy_call): Expand LEN, STACK_OFFSET to LONGEST.
	Use plongest to format print LEN.
	* mn10300-tdep.c (mn10300_push_dummy_call): Expand LEN to
	LONGEST.
	* mt-tdep.c (mt_push_dummy_call): Expand STACK_DEST, TYPELEN to
	LONGEST.
	* ppc-sysv-tdep.c (ppc_sysv_abi_push_dummy_call): Expand
	ARGSPACE, ARGOFFSET, STRUCTOFFSET, LEN, to LONGEST.
	(ppc64_sysv_abi_push_dummy_call): Expand BYTE, LEN to ssize_t.
	* rl78-tdep.c (rl78_push_dummy_call): Expand LEN, CONTAINER_LEN
	to LONGEST.
	* rs6000-aix-tdep.c (rs6000_push_dummy_call): Expand ARGBYTES,
	LEN to LONGEST.
	(ran_out_of_registers_for_arguments): Expand SPACE, JJ to
	LONGEST.
	* s390-tdep.c (s390_value_from_register): Expand LEN to LONGEST.
	(is_power_of_two): Expand parameter N to ULONGEST.
	(s390_push_dummy_call): Expand LENGTH to ULONGEST.
	* score-tdep.c (score_push_dummy_call): Expand ARGLEN to
	LONGEST.
	* sh-tdep.c (sh_use_struct_convention) Expand len to LONGEST.
	(sh_justify_value_in_reg): Expand parameter LEN to LONGEST.
	(sh_push_dummy_call_fpu): Expand LEN to LONGEST.  Expand REG_SIZE
	to ssize_t.
	(sh_push_dummy_call_nofpu): Likewise.
	* sh64-tdep.c (sh64_push_dummy_call): Expand STACK_OFFSET,
	STACK_ALLOC, LEN to LONGEST.
	* sparc-tdep.c (sparc32_store_arguments): Expand LEN to LONGEST.
	* sparc64-tdep.c (sparc64_store_floating_fields): Expand
	parameter BITPOS to LONGEST.  Expand SUBPOS to LONGEST.
	(sparc64_extract_floating_fields): Likewise.
	(sparc64_store_arguments): Expand LEN to LONGEST.
	* spu-tdep.c (spu_push_dummy_call): Expand N_REGS, LEN to LONGEST.
	(spu_value_from_register): Expand LEN to LONGEST.
	* tic6x-tdep.c (tic6x_push_dummy_call): Expand REFERENCES_OFFST,
	LEN to LONGEST.  Expand LEN to ssize_t.  Use plongest to format
	print LEN.
	* tilegx-tdep.c (tilegx_push_dummy_call): Expand TYPELEN,
	SLACKLEN, ALIGNLEN to LONGEST.
	* v850-tdep.c (v850_push_dummy_call): Expand LEN to LONGEST.
	* vax-tdep.c (vax_store_arguments): Expand COUNT, LEN to LONGEST.
	(vax_return_value): Expand LEN to LONGEST.
	* xstormy16-tdep.c (xstormy16_push_dummy_call): Expand J, TYPELEN
	to LONGEST.
	* xtensa-tdep.c (xtensa_store_return_value): Print LEN instead of
	TYPE_LENGTH..
	(struct argument_info): Expoand member LENGTH to ssize_t.
	(struct argument_info.u): Expand member OFFSET to ssize_t.
	(xtensa_push_dummy_call): Expand SIZE, ONSTACK_SIZE to LONGEST.
	Expand N to ssize_t.  Use pulongest to format print TYPE_LENGTH.

--MP_/X_WjDOvz/B_fvlsrmCwRdxe
Content-Type: text/x-patch
Content-Transfer-Encoding: 7bit
Content-Disposition: attachment; filename=bitpos-tdep.patch

+++ gdb/avr-tdep.c
@@ -1170,13 +1170,14 @@ avr_dummy_id (struct gdbarch *gdbarch, struct frame_info *this_frame)
 
 struct stack_item
 {
-  int len;
+  ssize_t len;
   struct stack_item *prev;
   void *data;
 };
 
 static struct stack_item *
-push_stack_item (struct stack_item *prev, const bfd_byte *contents, int len)
+push_stack_item (struct stack_item *prev, const bfd_byte *contents,
+		 ssize_t len)
 {
   struct stack_item *si;
   si = xmalloc (sizeof (struct stack_item));
@@ -1265,12 +1266,12 @@ avr_push_dummy_call (struct gdbarch *gdbarch, struct value *function,
 
   for (i = 0; i < nargs; i++)
     {
-      int last_regnum;
-      int j;
+      ssize_t last_regnum;
+      ssize_t j;
       struct value *arg = args[i];
       struct type *type = check_typedef (value_type (arg));
       const bfd_byte *contents = value_contents (arg);
-      int len = TYPE_LENGTH (type);
+      ssize_t len = TYPE_LENGTH (type);
 
       /* Calculate the potential last register needed.  */
       last_regnum = regnum - (len + (len & 1));
